repo-tools: refactor api report generation
